Compatibility with Akka 2.5.0
