+++ .travis.yml
@@ -26,14 +26,14 @@ env:
   - CI_CATEGORY=UNIT_TESTS
   - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.3.13 ITERATEES_VERSION=2.3.8 MONGO_PROFILE=default MONGO_VER=2_6
   - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.3.13 ITERATEES_VERSION=2.3.8 MONGO_PROFILE=default MONGO_VER=3
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
-  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
+  - CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
 matrix: 
-  exclude: 
+  exclude:
     - scala: 2.10.5
       jdk: oraclejdk8
       env: CI_CATEGORY=UNIT_TESTS
@@ -52,29 +52,29 @@ matrix:
     - jdk: oraclejdk8
       env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.3.13 ITERATEES_VERSION=2.3.8 MONGO_PROFILE=default MONGO_VER=3
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=default MONGO_VER=3_4
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=invalid-ssl MONGO_VER=3_4
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=mutual-ssl MONGO_VER=3_4
     - scala: 2.10.5
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
     - jdk: oraclejdk7
-      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.4.17 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
+      env: CI_CATEGORY=INTEGRATION_TESTS AKKA_VERSION=2.5.0 ITERATEES_VERSION=2.6.1 MONGO_PROFILE=rs MONGO_VER=3_4
 before_install:
 - ./.ci_scripts/beforeInstall.sh $CI_CATEGORY $MONGO_VER $MONGO_PROFILE
 - if [ "x$CI_CATEGORY" = "xINTEGRATION_TESTS" ]; then (./.ci_scripts/fork-mongod.sh); fi
