+++ bson/src/main/scala/bson.scala
@@ -52,6 +52,22 @@ object `package` extends DefaultBSONHandlers {
 
   /** Convenient type alias for document handlers */
   type BSONDocumentHandler[T] = BSONDocumentReader[T] with BSONDocumentWriter[T] with BSONHandler[BSONDocument, T]
+
+  /** Handler factory */
+  def BSONDocumentHandler[T](
+    read: BSONDocument => T,
+    write: T => BSONDocument
+  ): BSONDocumentHandler[T] = new BSONDocumentHandlerImpl[T](read, write)
+
+  private final class BSONDocumentHandlerImpl[T](
+    r: BSONDocument => T,
+    w: T => BSONDocument
+  ) extends BSONDocumentReader[T]
+      with BSONDocumentWriter[T] with BSONHandler[BSONDocument, T] {
+
+    def read(doc: BSONDocument): T = r(doc)
+    def write(value: T): BSONDocument = w(value)
+  }
 }
 
 object BSON {
