+++ project/plugins.sbt
@@ -8,6 +8,8 @@ resolvers ++= Seq(
    "https://dl.bintray.com/typesafe/sbt-plugins/"))(Resolver.ivyStylePatterns),
   Resolver.url("eed3si9n-repo", url(
     "https://dl.bintray.com/eed3si9n/sbt-plugins/"))(Resolver.ivyStylePatterns),
+  Resolver.url("typesafe-repo", url(
+   "https://dl.bintray.com/typesafe/sbt-plugins/"))(Resolver.ivyStylePatterns),
   Resolver.url("jsuereth-repo", url(
     "https://dl.bintray.com/jsuereth/sbt-plugins/"))(Resolver.ivyStylePatterns))
